[Chore] DateType 변환 유틸함수 분리
