+++ src/components/list/ReservationOption.tsx
@@ -5,6 +5,7 @@ import Image from 'next/image';
 import { OptionContext } from '@/context/OptionContext';
 import { useContext } from 'react';
 import { useOption } from '@/hooks/useOption';
+import { formatDate } from '@/utils/formatDate';
 export interface DateType {
   year: number;
   month: number;
@@ -21,14 +22,6 @@ function ReservationOption() {
   const { selectedDate } = useOption();
   const { headCount, modifyHeadCount, open } = useContext(OptionContext);
 
-  const formatDate = (date: DateType) => {
-    const { year, month, day } = date;
-    const formatted_year = String(year);
-    const formatted_month = String(month).padStart(2, '0');
-    const formatted_day = String(day).padStart(2, '0');
-    return [formatted_year, formatted_month, formatted_day];
-  };
-
   const onClick = (event: React.MouseEvent<HTMLDivElement>) => {
     const target = event.target as HTMLImageElement;
     if (target.closest('button') === null) return;
@@ -38,16 +31,16 @@ function ReservationOption() {
 
   return (
     <div className="mb-2.5 flex gap-2" onClick={onClick}>
-      <button className="flex w-full items-center rounded-lg border border-GRAY_200 py-2 pl-2 outline-none">
+      <button className="flex items-center w-full py-2 pl-2 border rounded-lg outline-none border-GRAY_200">
         <Image src={calendar} alt="calendar icon" className="mr-2" />
-        <p className="font-system5 text-system5 text-GRAY_900">
+        <p className="text-system5 font-system5 text-GRAY_900">
           {formatDate(selectedDate || initialDate).join('-')}
         </p>
       </button>
 
-      <button className="flex w-full items-center rounded-lg border border-GRAY_200 p-2 outline-none">
+      <button className="flex items-center w-full p-2 border rounded-lg outline-none border-GRAY_200">
         <Image src={person_icon} alt="person icon" className="mr-2" />
-        <div className="flex w-full justify-between">
+        <div className="flex justify-between w-full">
           <p
             className={`text-system5 text-GRAY_900 ${
               headCount ? 'font-system5_bold' : 'font-system5'
